cherry pick Kuba's commit
